modularize expr_gen
